# users-service/docs/swagger.yaml
basePath: /
host: localhost:3001
info:
  contact: {}
  description: Microservice for students, groups, teachers and staff
  title: Users Service API
  version: "1.0"
paths:
...
  /groups/{id}:
    get:
      description: Get group by ID
      parameters:
      - description: group id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get group
      tags:
      - Groups
...
  /students/{id}:
    get:
      description: Get student by ID
      parameters:
      - description: student id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get student
      tags:
      - Students
...
  /teachers/{id}:
    get:
      description: Get teacher by ID
      parameters:
      - description: teacher id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get teacher
      tags:
      - Teachers
swagger: "2.0"